% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wapqe.R
\name{wapqe}
\alias{wapqe}
\title{Rao’s apportionment of the quadratic entropy (APQE).}
\usage{
wapqe(df, dis = NULL, structures = NULL, formula = c("QE", "EDI"),
  wopt = c("even", "speciesab"), tol = 1e-08)
}
\arguments{
\item{df}{Dataframe or matrix with sites as rows and species as columns. Entries are abundances of species within sites.}

\item{dis}{Dissimilarity among species (NULL or class 'dist').}

\item{structures}{Data frame that contains the name of the group (row) of an level (column) to which the site belongs. Sites in structures should be in the same order as in df. Default is NULL.}

\item{formula}{Quadratic entropy formula ("QE", "EDI"). "QE" is default.}

\item{wopt}{Site weighting type ("even", "speciesab"). Default is "even".}

\item{tol}{A tolerance threshold (a value less than tol is considered equal to zero).}
}
\value{
A data frame with each component of the selected diversity decomposition.
}
\description{
The hierarchical apportionment of quadratic entropy defined by Rao (1982) wich allows biodiversity to be considered in terms of equivalent number of species, sites, regions, etc., in a nested hierarchy of multiple scales.
}
\details{
For the associated permutation test see \code{\link{randtestapqe}}.
}
\examples{
data(macroloire)
# Taxonomic dissimilarities among species:
dTaxo <- dist.taxo(macroloire$taxo)^2/2
dTaxo <- dTaxo/max(dTaxo)
# Size-based dissimilarities among species
dSize <- dist.prop(macroloire$traits[ ,1:4], method = 2)
# Dissimilarities among species in terms of feeding categories
dFeed <- dist.prop(macroloire$traits[ ,5:11], method = 2)
# Dissimilarities among species in terms of both size and feeding categories
dSF <- (dSize+dFeed)/2

# Table with sites as rows (stations), species as columns and abundances as entries
ab <- as.data.frame(t(macroloire$fau))
# Table with sites as rows and one column only. Entries indicate the geological region associated with each site
stru <- macroloire$envir["Morphoregion"]

wapqe(ab, , stru, formula = "QE")
wapqe(ab, dTaxo, stru, formula = "QE")
wapqe(ab, dSize, stru, formula = "QE")
wapqe(ab, dFeed, stru, formula = "QE")
wapqe(ab, dSF, stru, formula = "QE")

}
\author{
Sandrine Pavoine, Eric Marcon, Carlo Ricotta.
}
\references{
Pavoine, S., Marcon, E. and Ricotta, C. (2016), ‘Equivalent numbers’ for species, phylogenetic or functional diversity in a nested hierarchy of multiple scales. Methods Ecol Evol, 7: 1152–1163. DOI:10.1111/2041-210X.12591
Rao, C.R. (1982) Diversity: its measurement, decomposition, apportionment and analysis. Sankhya: The Indian Journal of Statistics, A44, 1–22.
Pavoine S. and Dolédec S. (2005) The apportionment of quadratic entropy: a useful alternative for partitioning diversity in ecological data. Environmental and Ecological Statistics, 12, 125–138. DOI:10.1007/s10651-005-1037-2
}
\seealso{
\code{\link{randtestEqRSintra}}, \code{\link{EqRS}}, \code{\link{EqRao}}, \code{\link{apqe}} in ade4.
}

